Adding the JSON formats of generic models (Name, Title, ... etc) and changing the handling of double properties in postgres to be float

+++ lampetia-postgresql/src/main/scala/lampetia/sql/dialect/postgresql/Postgresql.scala
@@ -24,7 +24,7 @@ trait Postgresql
     def name(propertyType: PropertyType[_]): String = propertyType match {
       case IntProperty => "integer"
       case FloatProperty => "float"
-      case DoubleProperty => "double"
+      case DoubleProperty => "float"
       case LongProperty => "long"
       case StringProperty => "varchar"
       case DateProperty  => "date"
